Refactoring to edu.purdue.sigbots.cli.management.*
Fixes T5
